Update jinkfsutil.c
Fixing the 'Reserved : 0' bug in the Disk information.
Replacing uint8_t by a uint16_t (according to the JinkFS standard)

+++ tools/jinkfsutil.c
@@ -19,7 +19,7 @@ typedef struct {
     uint8_t Signature[3];
     uint8_t DiskLabel[8];
     uint16_t BytesPerBlock;
-    uint8_t Reserved;
+    uint16_t Reserved;
     uint32_t FileTableOffset;
     uint32_t BlockAreaOffset;
     uint8_t TotalOfEntries;
